Seperate updateable from drawable
